test: add unit test cases for vchart event
